feat: Implement Head of Family management
- Created HeadOfFamilyStoreRequest for validation during creation.
- Created HeadOfFamilyUpdateRequest for validation during updates.
- Added HeadOfFamilyResource for API responses.
- Defined HeadOfFamilyRepositoryInterface for repository pattern.
- Implemented HeadOfFamilyRepository for data handling.
- Added search scope in HeadOfFamily model for filtering.
- Registered HeadOfFamilyRepository in RepositoryServiceProvider.
- Created HeadOfFamilyFactory for seeding.
- Updated migration for head_of_families table to fix identity_number type.
- Added migration to rename inditity_number to identity_number.
- Created HeadOfFamilySeeder for populating test data.
- Updated DatabaseSeeder to include HeadOfFamilySeeder.
- Added English and Indonesian language files for authentication, pagination, passwords, and validation.
- Updated API routes to include head-of-family resource.

+++ Backend/database/migrations/2025_09_12_075828_create_head_of_families_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->string('profile_picture')->nullable();
-            $table->integer('inditity_number')->unique();
+            $table->bigInteger('inditity_number')->unique();
             $table->enum('gender', ['male', 'female']);
             $table->date('birth_date');
             $table->string('phone_number')->nullable();
